Change to be able to upload images when adding a pin

+++ lib/model/firebase_api.dart
@@ -13,7 +13,7 @@ class FirebaseApi {
   FirebaseApi._internal();
 
   factory FirebaseApi() => instance;
-  final storageRef = FirebaseStorage.instance.ref();
+  final storage = FirebaseStorage.instance.ref();
   final db = FirebaseFirestore.instance;
 
   Future<String> addToFirestore({
@@ -24,8 +24,8 @@ class FirebaseApi {
       final document = await db.collection(root).add(data);
       _log("Add document to id: ${document.id}");
       return document.id;
-    } catch (e) {
-      _errorLogger(e);
+    } catch (e, stackTrace) {
+      _errorLogger(e, stackTrace);
       return "";
     }
   }
@@ -43,8 +43,8 @@ class FirebaseApi {
         await db.collection(root).doc(path).set(data);
       }
       _log("Write document: $path");
-    } catch (e) {
-      _errorLogger(e);
+    } catch (e, stackTrace) {
+      _errorLogger(e, stackTrace);
     }
   }
 
@@ -62,58 +62,58 @@ class FirebaseApi {
       }
       _log("Fetch pins: $pins");
       return pins;
-    } catch (e) {
-      _errorLogger(e);
+    } catch (e, stackTrace) {
+      _errorLogger(e, stackTrace);
       return [];
     }
   }
 
   Future<String?> getStoragePath({required String firestoreId}) async {
     if (firestoreId.isEmpty) {
-      return null;
+      return "Temp";
     }
     try {
       final docRef = db.collection("storage").doc(firestoreId);
       final document = await docRef.get();
       final Map<String, dynamic>? data = document.data();
       final path = data?["path"];
-      if (path != null) {
-        _log("Get storage path: $path");
-        return path;
-      }
-      final generatedPath = const Uuid().v4();
-      writeToFirestore(
-        root: "storage",
-        path: firestoreId,
-        data: {
-          "path": generatedPath,
-        },
-        update: true,
-      );
-      _log("Get storage path: $generatedPath");
-      return generatedPath;
-    } catch (e) {
-      _errorLogger(e);
+      _log("Get storage path: $path");
+      return path;
+    } catch (e, stackTrace) {
+      _errorLogger(e, stackTrace);
       return null;
     }
   }
 
   Future<bool> uploadToStorage({
-    required String path,
+    required String firestoreId,
+    required String name,
     required File data,
   }) async {
     try {
-      await storageRef.child(path).putFile(data);
+      String? path = await getStoragePath(firestoreId: firestoreId);
+      if (path == null) {
+        path = const Uuid().v4();
+        await writeToFirestore(
+          root: "storage",
+          path: firestoreId,
+          data: {
+            "path": path,
+          },
+          update: true,
+        );
+      }
+      await storage.child("$path/$name").putFile(data);
       _log("Upload to storage: $path");
       return true;
-    } catch (e) {
-      _errorLogger(e);
+    } catch (e, stackTrace) {
+      _errorLogger(e, stackTrace);
       return false;
     }
   }
 
   Future<List<Reference>> fetchStorageRefs({required String path}) async {
-    final fileList = await storageRef.child(path).listAll();
+    final fileList = await storage.child(path).listAll();
     final sortedFileList = fileList.items.reversed.toList();
     _log("Fetch storage refs: $sortedFileList");
     return sortedFileList;
@@ -123,16 +123,56 @@ class FirebaseApi {
     try {
       final deletesAsync = <Future>[];
       final storagePath = await getStoragePath(firestoreId: firestoreId);
-      final storageRefs = await fetchStorageRefs(path: storagePath!);
-      for (final ref in storageRefs) {
-        deletesAsync.add(ref.delete());
+      if (storagePath != null) {
+        final storageRefs = await fetchStorageRefs(path: storagePath);
+        for (final ref in storageRefs) {
+          deletesAsync.add(ref.delete());
+        }
       }
-
       deletesAsync.add(db.collection("storage").doc(firestoreId).delete());
       await Future.wait(deletesAsync);
       _log("Delete from storage to id: $firestoreId");
-    } catch (e) {
-      _errorLogger(e);
+    } catch (e, stackTrace) {
+      _errorLogger(e, stackTrace);
+    }
+  }
+
+  Future<void> deleteTempFiles() async {
+    try {
+      final deletesAsync = <Future>[];
+      final storageRefs = await fetchStorageRefs(path: "Temp");
+      for (final ref in storageRefs) {
+        deletesAsync.add(ref.delete());
+      }
+      await Future.wait(deletesAsync);
+      _log("Delete temp files from storage");
+    } catch (e, stackTrace) {
+      _errorLogger(e, stackTrace);
+    }
+  }
+
+  Future<void> copyStorageFiles({
+    required String srcPath,
+    required String destPath,
+  }) async {
+    try {
+      final srcAsync = <Future>[];
+      final storageRefs = await fetchStorageRefs(path: srcPath);
+      for (final ref in storageRefs) {
+        srcAsync.add(ref.getData());
+      }
+      final srcData = await Future.wait(srcAsync);
+
+      final destAsync = <Future>[];
+      for (int i = 0; i < storageRefs.length; i++) {
+        destAsync.add(
+          storage.child("$destPath/${storageRefs[i].name}").putData(srcData[i]),
+        );
+      }
+      await Future.wait(destAsync);
+      _log("Copy storage files: $storageRefs");
+    } catch (e, stackTrace) {
+      _errorLogger(e, stackTrace);
     }
   }
 
@@ -142,9 +182,10 @@ class FirebaseApi {
     }
   }
 
-  void _errorLogger(dynamic e) {
+  void _errorLogger(dynamic e, dynamic stackTrace) {
     if (kDebugMode) {
       print(e);
+      print(stackTrace);
     }
     SnackBarHelper().show("エラー: $e");
   }
